requirements.txt
vkbottle
Pillow

